// src/proto/user-settings.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.11.5
//   protoc               v6.33.6
// source: src/proto/user-settings.proto

/* eslint-disable */
import type { Metadata } from '@grpc/grpc-js';
import { GrpcMethod, GrpcStreamMethod } from '@nestjs/microservices';
import { Observable } from 'rxjs';

export const protobufPackage = 'settings';

export interface TfaRequest {
  id: string;
}

export interface TfaVerifyRequest {
  id: string;
  secret: string;
}

export interface TfaResponse {
  status: TfaResponse_Status;
  message?: string | undefined;
}

export enum TfaResponse_Status {
  FAIL = 0,
  OK = 1,
  UNRECOGNIZED = -1,
}

export interface AdaptiveRequest {
  id: string;
}

export interface AdaptiveResponse {
  active: boolean;
}

export const SETTINGS_PACKAGE_NAME = 'settings';

export interface SettingsProxyServiceClient {
  reject2Fa(request: TfaRequest, metadata?: Metadata): Observable<TfaResponse>;

  verify2Fa(request: TfaVerifyRequest, metadata?: Metadata): Observable<TfaResponse>;

  acceptAdaptive(request: AdaptiveRequest, metadata?: Metadata): Observable<AdaptiveResponse>;
}

export interface SettingsProxyServiceController {
  reject2Fa(request: TfaRequest, metadata?: Metadata): Promise<TfaResponse> | Observable<TfaResponse> | TfaResponse;

  verify2Fa(
    request: TfaVerifyRequest,
    metadata?: Metadata,
  ): Promise<TfaResponse> | Observable<TfaResponse> | TfaResponse;

  acceptAdaptive(
    request: AdaptiveRequest,
    metadata?: Metadata,
  ): Promise<AdaptiveResponse> | Observable<AdaptiveResponse> | AdaptiveResponse;
}

export function SettingsProxyServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = ['reject2Fa', 'verify2Fa', 'acceptAdaptive'];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcMethod('SettingsProxyService', method)(constructor.prototype[method], method, descriptor);
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcStreamMethod('SettingsProxyService', method)(constructor.prototype[method], method, descriptor);
    }
  };
}

export const SETTINGS_PROXY_SERVICE_NAME = 'SettingsProxyService';